% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knobi_env.R
\name{knobi_env}
\alias{knobi_env}
\title{KBPM environmental analysis}
\usage{
knobi_env(
  knobi_results,
  data,
  control = NULL,
  plot_out = FALSE,
  plot_filename = NULL,
  plot_dir = NULL
)
}
\arguments{
\item{knobi_results}{The output object of \code{\link{knobi_fit}} function (main package function).}

\item{data}{A list containing the following data: \itemize{
\item env: data frame containing the values of every environmental variable(s) in each column. Rows represent years.
\item years: years in which the environmental variable(s) are reported.}}

\item{control}{Optional. List containing the following settings: \itemize{
\item nlag: this argument is used to test, in the correlation analysis, the lags smaller or equal to 'nlag' (natural number). This means that correlation between KBPM residuals_{t} and X_{t-lag}, being X the environmental variable and lag the selected value from sequence {0,1,...,nlag}, is computed. The lag corresponding to the highest correlation among the KBPM residuals and the corresponding time lagged environmental covariable is considered in the environmental model (unless otherwise specified in 'selected_var'). By default, 'nlag=3'. See details.
\item lag: optional numerical vector providing the lag value(s) to consider in the relation between the KBPM surplus production and the environmental variable(s). The length of this argument must be equal to the number of environmental variables included. Applies only if 'nlag' argument is not provided.
\item start_c: optional. Numerical vector providing the starting values of the environmental parameter 'c' for the additive and multiplicative models, respectively. By default, start_c=c(1,1). See details.
\item ar_cor: optional. Logical. By default this argument is FALSE, meaning that the correlation between the KBPM residuals and the environmental variable(s) is analyzed through Pearson correlation test, as described above. If this argument is "TRUE", the relationship  between the KBPM residuals and the environmental variables is analyzed by fitting autoregressive (AR) models including each one of the environmental time lagged variables as explanatory covariables. The environmental variable whose model reports the lowest Akaike information criterion (AIC) is selected to be included in the environmental KBPM fit. See details.
\item plot3d: optional. Logical. If this argument is TRUE, 3D plots reporting the surplus production curve conditioned to a grid of environmental values are provided. FALSE by default.
\item selected_var: optional. Character. By default, the fit is carried out  including the highest correlated environmental variable at the corresponding time lag. However, if this argument is equal to the name of one of the environmental variables, this variable is used in the environmental fit considering the lag derived from the correlation analysis.
\item multicovar: optional. Logical. TRUE  means that the environmental model includes all the input environmental covariables at the same time, up to a maximum of 5. By default this argument is FALSE, which means that only the environmental variable reporting the highest correlation is included (after lagging it if corresponds).}}

\item{plot_out}{Logical. TRUE means that a file with the plot of the environmental fits is created. The default value is the input in the \code{\link{knobi_fit}} function.}

\item{plot_filename}{Optional. Name of the folder that will contain the plots. Required when 'plot_out=TRUE'. The default value is the input in the \code{\link{knobi_fit}} function.}

\item{plot_dir}{Optional. Directory to create the folder and save the plots. Required when 'plot_out=TRUE'. The default value is the input in the \code{\link{knobi_fit}} function.}
}
\value{
A list containing the results of the three-step environmental analysis is provided. \itemize{
\item selected_lag: data frame providing the time lag of the environmental variable(s) in the KBPM fit.
\item lag_cor: correlation between the environmental variable(s) and the KBPM residuals for each one of the time lags.
\item env_aic: if 'ar_cor=TRUE', AIC values of each one of the autoregressive models (see details).
\item selected_var: environmental variable used in the fit.
\item model_env_Multiplicative: estimates of the multiplicative model parameters.
\item model_env_Additive: estimates of the additive model parameters.
\item ref_pts: reference points (RPs) estimates for each model assuming X_t=0 (see vignettes for RPs equations depending on X_t).
\item scaled_environmental_var: standardized environmental variable(s) used in the fit, i.e. the environmental variable(s) minus its mean and divided by its standard deviation (sd).
\item plots3D: list with the 3D plots objects (if 'plot3d=TRUE').
\item goodness_of_fit: list of goodness-of-fit measures of the fits: \itemize{
\item residuals: Pearson residuals from each model fit (base KBPM, additive model and multiplicative model).
\item error_table: array of performance and accuracy measures for each model: Standard error of the regression (SER), coefficient of determination (R-squared), adjusted coefficient of determination (adj-R-squared), Akaike information criterion (AIC), root-mean-squared error (RMSE) and mean absolute percentage error (MAPE) of observed and estimated surplus production values; and the statistic and p-value of the F-test. p-values lower than the considered significance level lead to the rejection of the null hypothesis, which implies that the KBPM environmental model significantly improves the KBPM model that does not consider environmental information (H0: there is no improvement of the environmental model in relation to the base model).}}
Results are also reported through plots which are displayed in the plot window and also saved (if 'plot_out=TRUE') in the provided directory or in the same directory as knobi_fit's plots.
The first plot reports the correlation analysis between the environmental variable(s) and the KBPM residuals. The second one reports the fitted SP values derived from the base model (no environmental information) and from the environmental ones.
If 'multicovar=FALSE' and 'plot3d=TRUE', 3D plots reporting the surplus production curve conditioned to a grid of environmental values are also reported.
If 'multicovar=TRUE', a plot with the Pearson correlation between the environmental variables is reported too.
}
\description{
Analyze and model the relationships between surplus production (SP) and environmental covariable(s) to test whether productivity changes in response to environmental fluctuations in three steps: (1) correlation analysis between the environmental variable(s) at different delays (lags) and the KBPM residuals through Pearson's correlation or autoregressive models; (2) selection of which lagged environmental variable(s) is included in the environmental KBPM models fit; (3) KBPM environmental fitting, where environmental effects are included as additive and multiplicative effects in the KBPM formulation (see details).
}
\details{
Additive environmental model adds the following term on the right hand of equation (1) or (2) described in \code{\link{knobi_fit}} function: \eqn{cX_{t-lag}B_{t}}, being \eqn{X_{t-lag}} the environmental variable and \eqn{B_{t}} the biomass or SSB at time \eqn{t-lag}.
Multiplicative environmental model multiplies the right hand of equation (1) or (2) by \eqn{exp(cX_{t-lag})}.

If ar_cor argument is "TRUE", the correlation analysis between the KBPM residuals and the environmental variable(s) is carried out as follows.
First, an AR model is fitted to the residuals.
\deqn{r_t=\sum_{i=1}^{p}\beta_{i}r_{t-i}+\epsilon_{t}}
being \eqn{r_t} the KBPM base residual for year \eqn{t} and \eqn{p} the AR model order, estimated as the maximum time lag at which the absolute value of the residuals partial autocorrelation is large than \eqn{qnorm(0.975)/\sqrt(N_r)}, being \eqn{N_r} the length of the residuals series.
Then, AR models are fitted considering each one of the lagged environmental variable(s),
\deqn{r_{t,lag}=\sum_{i=1}^{p}\beta_{i}r_{t-i}+X_{t-lag}+\epsilon_{t}}, for \eqn{lag=0,1,...,nlag}.
being \eqn{X_{t,lag}} the lagged environmental variable at year {t-lag}. Then, we have an autoregressive model for each of the lagged environmental variables.
The AIC values of the above models are compared, and the lagged environmental variable whose model reports the lowest AIC is used in the KBPM fit, except if the argument 'lag' is used.
}
\examples{

\dontrun{

# First, run the example of knobi_fit function

# Then, provide environmental data series

Env<-data.frame(years=seq(1973,2020),AMO=c(-0.236,-0.441,-0.32,-0.385,-0.21,-0.201,-0.132,
-0.041,-0.098,-0.235,-0.093,-0.23,-0.29,-0.297,0.044,-0.028,-0.106,-0.061,-0.155,-0.242,
-0.234,-0.2,0.112,-0.082,0.028,0.349,0.094,0.004,0.095,0.041,0.207,0.182,0.268,0.242,
0.123,0.114,0.015,0.325,0.078,0.189,0.142,0.077,0.09,0.318,0.291,0.045,0.15,0.279),
TMax_Vigo=c(16.5,16.7,17.1,16.3,16.4,16.7,17.3,17.3,17.4,18,17.5,17.5,17.6,18,17,18.4,
17.8,19.6,19.1,18,17.9,17.7,17.7,19.4,18.2,19.7,19.2,18.6,18,18.3,18.5,18.8,18.7,18.7,
18.5,17.9,17.4,19.2,19,19.7,18,19.1,19.4,20,19.5,20.2,18.8,18.6))

# The environmental data series must start in the first year of the KBPM fit data
# minus the provided nlag or lag
years<-knobi_results$data$years # See knobi_fit example to obtain the knobi_results object
ind<-which(Env[,1]==years[1])
ind1<-which(Env[,1]==years[length(years)])
nlag<-5
Env<-Env[(ind-nlag):ind1,]

# Now we create the environmental list
data<-list(env=data.frame(AMO=Env$AMO,Tmax=Env$TMax_Vigo),
           years=Env$years)
control<-list(nlag=nlag)

knobi_environmental<-knobi_env(knobi_results,data,control)
knobi_environmental
}

}
\author{
\itemize{
\item{Anxo Paz}
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{M. Grazia Pennino}
}
}
