% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knobi_env.R
\name{knobi_env}
\alias{knobi_env}
\title{KBPM environmental analysis}
\usage{
knobi_env(
  knobi_results,
  data,
  control = NULL,
  plot_out = FALSE,
  plot_filename = NULL,
  plot_dir = NULL
)
}
\arguments{
\item{knobi_results}{The output object of \code{\link{knobi_fit}} function (main function).}

\item{data}{A list containing the following data: \itemize{
\item env: data frame containing the values of each one of the environmental variable(s) in one column. Each row represents a year.
\item years: time series of years corresponding to the environmental variable(s).}}

\item{control}{A list containing the following settings: \itemize{
\item nlag: this argument is used to test all the lags smaller or equal to nlag (numeric) through cor.test function. The lag corresponding to the highest pearson correlation among the base KBPM surplus production residuals and the lagged environmental covariable values is considered in the environmental model. By default, 'nlag=3'.
\item lag: optional numerical vector providing the used lag value(s) in the relation among the base KBPM surplus production residuals and the environmental variable(s). This means that the residuals_{t} is related to X_{t-lag} being X the environmental variable. The length of this argument must be equal to the number of environmental variables included. Only work if 'nlag' argument is not provided.
\item start_c: optional numerical vector providing the start values of the environmental c parameter for the optimization of the additive and multiplicative models, respectively. By default, start_c=c(1,1). See details.
\item ar_cor: optional logical. By default this argument is FALSE, which means the correlation between the KBPM base residuals and the environmental variable(s) is tested through a pearson correlation test, as mentioned above. If this argument is "TRUE", the correlation is estimated by fitting an autoregressive (AR) model for the KBPM residuals and compare its Akaike information criterion (AIC) with the AIC obtained for the same model considering the environmental variable as a regressive variable. See details.
\item plot3d: optional logical. If this argument is TRUE, 3D plots reporting the surplus production curve conditioned to a grid of environmental values. FALSE by default.
\item selected_var: optional character. By default, the fit is done using the higher correlated environmental variable. If this argument is provided, the fit is made with this fixed selected variable considering its lags based on the 'nlag' or 'lag' arguments.
\item multicovar: optional logical. TRUE if you want to fit the environmental model including all the input environmental covariables, up to a maximum of 5. By default this argument is FALSE, which means that only the environmental covariable reporting the highest pearson correlation is included (after lagging it if corresponds).}}

\item{plot_out}{Logical. TRUE means that a file with the  environmental fit plots is created. By default this argument is FALSE.}

\item{plot_filename}{Optional name of the folder that will contain the plots. Required when plot_out=TRUE. The default value is the input of this argument in the knobi_fit function.}

\item{plot_dir}{Optional directory for creating the folder and save the plots. Required when plot_out=TRUE. The default value is the input of this argument in the knobi_fit function.}
}
\value{
A list containing the environmental analysis is provided. \itemize{
\item selected_lag: Data frame with the estimated lag corresponding to the one reporting the highest correlation (or the lowest AIC if 'ar_cor=TRUE') between the environmental variable and the base KBPM surplus production residuals (derived if lag is not fixed) or the fixed lag and the correlation corresponding to this lag for each variable.
\item lag_cor: Correlation between the environmental variable(s) value and the base KBPM surplus production residuals for each lag.
\item env_aic: If 'ar_cor=TRUE', AIC values obtained for the AR model considering each environmental variable(s) for each lag among the AIC value for the AR model considering only the KBPM base SP residuals.
\item selected_var: Environmental variable used in the fit, chosen by the user or the one derived from the highest pearson correlation procedure. In case that argument 'multicovar' is omitted, 'NULL' or equal to 'FALSE'.
\item model_env_Multiplicative: Estimates of the multiplicative model parameters.
\item model_env_Additive: Estimates of the additive model parameters.
\item ref_pts: Reference points (RPs) estimates for each model assuming X_t=0 (see vignettes).
\item scaled_environmental_var: Standardized variable used in the fit, with the 'scale' and 'center' attributes.
\item environmental_variables: Standardized covariables used in the fit (if 'multicovar=TRUE'), with the 'scale' and 'center' attributes.
\item plots3D: List with the 3D plots objects (if 'plot3d=TRUE').
\item error: List of performance and accuracy: \itemize{
\item residuals: Pearson's residuals from the fit calculated as (observations-estimates)/sd(observations) for each model (base KBPM, additive model and multiplicative model).
\item error_table: Array of performance and accuracy (observed vs. estimated) measures for each model: Standard error of the regression (SER), coefficient of determination (R-squared), adjusted coefficient of determination (adj-R-squared), Akaike information criterion (AIC), root-mean-squared error (RMSE), mean absolute percentage error (MAPE) and the value of the F statistic corresponding to the comparison of each environmental model respect to the base model (F-value) and its corresponding p-value (Pr(>F)).}}
Result plots are shown in the plot window and also saved (if 'plot_out=TRUE') on the provided directory or in the same directory as knobi_fit.
The first plot reports the correlation analysis between the environmental variable(s) and the KBPM SP residuals. The second one reports the fitted values of the base model (no environmental information) and of the environmental ones.
If 'multicovar=FALSE' and 'plot3d=TRUE', 3D plots reporting the surplus production curve conditioned to a grid of environmental values are also reported.
}
\description{
Analyse and model the relationships between surplus production and environmental covariables to test whether productivity changes in response to environmental fluctuations.  Environmental effects are included as additive and multiplicative effects in the general KBPM formulation (see details).
}
\details{
Additive environmental model adds the following term on the right hand of equation (1) or (2) detailed in \code{\link{knobi_fit}} function: \eqn{cX_{t}B_{t}}, being \eqn{X_{t}}the environmental variable and \eqn{B_{t}} the biomass or SSB at time \eqn{t}.
Multiplicative environmental model multiplies the right hand of equation (1) or (2) by \eqn{exp(cX_{t})}.
The subscript t denotes the time (years).

If ar_cor argument is "TRUE", for the correlation test between the KBPM residuals and the environmental variable(s):
First, an AR model is fitted for the residuals as follows:
\deqn{r_t=\sum_{i=1}^{p}\beta_{i}r_{t-i}+\epsilon_{t}}
being \eqn{r_t} the KBPM base SP residual for year \eqn{t} and \eqn{p} the AR model order, estimated as the maximum partial autocorrelation function value that is bigger than \eqn{qnorm(0.975)/\sqrt(length(r))} in absolute value.
Then, an AR model is fitted considering the lagged environmental variable(s) considering each variable and lag,
\deqn{r_t=\sum_{i=1}^{p}\beta_{i}r_{t-i}+X_{t,lag}+\epsilon_{t}}
Being \eqn{X_{t,lag}} the lagged environmental variable for each year {t} and for each lag corresponding to \eqn{lag=0,1,...,nlag}.
For each model, the AIC is estimated, comparing their values in order to see what model has a smaller AIC and its environmental variable with its corresponding lag is used in the environmental adjustment if the environmental variable is not fixed.
}
\examples{

\dontrun{

# First, run the example of knobi_fit function

# Then, provide environmental data series

Env<-data.frame(years=seq(1973,2020),AMO=c(-0.236,-0.441,-0.32,-0.385,-0.21,-0.201,-0.132,
-0.041,-0.098,-0.235,-0.093,-0.23,-0.29,-0.297,0.044,-0.028,-0.106,-0.061,-0.155,-0.242,
-0.234,-0.2,0.112,-0.082,0.028,0.349,0.094,0.004,0.095,0.041,0.207,0.182,0.268,0.242,
0.123,0.114,0.015,0.325,0.078,0.189,0.142,0.077,0.09,0.318,0.291,0.045,0.15,0.279),
TMax_Vigo=c(16.5,16.7,17.1,16.3,16.4,16.7,17.3,17.3,17.4,18,17.5,17.5,17.6,18,17,18.4,
17.8,19.6,19.1,18,17.9,17.7,17.7,19.4,18.2,19.7,19.2,18.6,18,18.3,18.5,18.8,18.7,18.7,
18.5,17.9,17.4,19.2,19,19.7,18,19.1,19.4,20,19.5,20.2,18.8,18.6))

# The environmental data series must start in the first year of the KBPM fit data
# minus the provided nlag or lag
years<-knobi_results$data$years # See knobi_fit example to obtain the knobi_results object
ind<-which(Env[,1]==years[1])
ind1<-which(Env[,1]==years[length(years)])
nlag<-5
Env<-Env[(ind-nlag):ind1,]

# Now we create the environmental list
data<-list(env=data.frame(AMO=Env$AMO,Tmax=Env$TMax_Vigo),
           years=Env$years)
control=list(nlag=nlag)

knobi_environmental<-knobi_env(knobi_results,data,control)
knobi_environmental
}

}
\author{
\itemize{
\item{Anxo Paz}
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{M. Grazia Pennino}
}
}
