% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knobi_fit.R
\name{knobi_fit}
\alias{knobi_fit}
\title{Known Biomass Production Model (KBPM) fit}
\usage{
knobi_fit(
  data,
  control = NULL,
  plot_out = FALSE,
  plot_filename = NULL,
  plot_dir = NULL
)
}
\arguments{
\item{data}{A list containing the data. \itemize{
\item Catch: catch time series observations.
\item Biomass: biomass time series estimated through a data-rich stock assessment model. If available, otherwise enter SSB in the next argument.
\item Spawning_Biomass: SSB time series estimated through a data-rich stock assessment model. If available, otherwise introduce biomass in the previous argument.
\item years: optional. Years in which catch observations has been measured.
\item Stock: optional. Character string with the stock name for using in the plot subtitles.
\item Recruitment: optional. Recruitment time series estimated through a data-rich stock assessment model. See details.
\item F_input: optional. Fishing mortality time series estimated through a data-rich stock assessment model. See details.
\item classF_input: optional. Character indicating the type of fishing mortality estimate. See details.
\item RP: optional. Estimates of the following biological reference points derived from the fit of a data-rich stock assessment model (see details): \itemize{
\item F_MSY: fishing mortality at Maximum Sustainable Yield (MSY).
\item B_MSY: biomass at MSY (or SSB depending on the value of the 'method' argument, see control list).
\item MSY: Maximum Sustainable Yield.
\item K: virgin biomass.}}}

\item{control}{A list containing the following control parameters. \itemize{
\item pella: Logical. TRUE means that Pella-Tomlinson model is used whereas FALSE means that Schaefer model is fitted (by default). See details.
\item start_r: optional. Start value of the growth rate model parameter r (i.e. intrinsic rate of natural increase). See details.
\item start_K: optional. Start value of the carrying capacity model parameter K (the maximum population size for growth to be positive). See details.
\item start_p: optional. Start value of the shape model parameter p. Used for Pella-Tomlinson model. See details.
\item method: sets whether the fit is carried using "SSB" or "Biomass". The argument is only required if both time series, 'Spawning_Biomass' and 'Biomass', are provided.}}

\item{plot_out}{Logical. TRUE means that files are created with the input time series plots and the fitting results. FALSE by default.}

\item{plot_filename}{Optional. Name of the folder that will contain the plot files. By default, "knobi_results". Required when plot_out=TRUE.}

\item{plot_dir}{Optional. Directory for saving the plot files. Required when plot_out=TRUE. Current directory by default.}
}
\value{
An output list updated with the fitting information. The control output is the input one updated with the information of the plot settings. The data output is also the updated input including the annual average biomass (mean of two consecutive years), in $data$Average_Biomass, the surplus production, in $data$SP, and the F estimates derive from knobi_fit, in $data$F_output. The results of the KBPM fit are in the $fit list slot, containing: \itemize{
\item Parameter_estimates: model parameters estimates.
\item data: the data used for the model.
\item RP: biological reference points estimates: \itemize{
\item K: virgin biomass.
\item B_MSY: biomass at maximum sustainable yield.
\item F_MSY: fishing mortality at maximum sustainable yield.
\item MSY: maximum sustainable yield.
\item MSYoverK: ratio of MSY and K. }
\item optimr: list of some results provided by \code{\link[optimr]{optimr}}: \itemize{
\item value: value of the function corresponding to the parameter estimation.
\item convergence: integer code. ‘0’ indicates successful completion in the optimization.
\item message: character string giving any additional information returned by the optimizer, or NULL.}
\item error: list of performance and accuracy measures: \itemize{
\item residuals: pearson's residuals calculated as (observations-estimates)/sd(observations).
\item error_table: data frame measures of estimates accuracy (error measures comparing observed an estimated values) and model performance: \itemize{
\item SER: standard error of the regression, calculated as the root of the rate between the sum of residual squares and the degrees of freedom of the regression.
\item R-squared: coefficient of determination.
\item adj-R-squared: adjusted coefficient of determination.
\item AIC: Akaike information criterion.
\item RMSE: root mean squared error.
\item MAPE: mean absolute percentage error.}}}
The plots are displayed in the plot window and also saved (if 'plot_out=TRUE') in the provided directory or in the current one (if 'plot_dir' is not provided). The following input quantities are plotted: fishing mortality time series, SSB, surplus production and catch time series. Plots of catch over fishing mortality, fishing mortality over SSB, and catch over SSB time series with a smooth line from a "loess" regression are also available. Plot of input-output time series of fishing mortality  is provided with horizontal lines at fishing mortality at MSY ( one line if input F_MSY is NULL). The analogous SSB plot is also reported. On the other hand, the fitted surplus production curve is plotted twice with the SSB and SP observations (first) and with the catch and SP observations (second).
}
\description{
This function, that is the main function of the knobi package, fits a type of surplus production models named known-biomass production models (KBPM) (MacCall, 2002). The surplus production curve is fitting using the catch time series and the biomass or SSB (Spawning Stock Biomass) estimates derived from the fit of a data-rich stock assessment model.
}
\details{


The KBPMs implemented in the current package are explained below.
Schaefer model (1):
\deqn{SP_{t} = r B_{t} (1-(B_{t}/K))}
where \eqn{SP_{t}} is the surplus production, \eqn{B_{t}} is the biomass or \eqn{SSB_{t}} averaged (mean of two consecutive years), \eqn{r} is the population growth rate parameter, and \eqn{K} is the virgin biomass. The subscript \eqn{t} denotes the time (years).
Pella and Tomlinson model (2):
\deqn{SP_{t} = (r/p) B_{t} (1-(B_{t}/K)^{p})}
where \eqn{SP_{t}} is the surplus production, \eqn{B_{t}} is the biomass or \eqn{SSB_{t}} averaged,  \eqn{r} is the population growth rate parameter, \eqn{K} is the virgin biomass and \eqn{p} is the asymmetry parameter. The subscript \eqn{t} denotes the time (years).

The recruitment values, F_input estimates and F_type argument are included to get an overview of the stock status, but are not needed for the KBPM fitting. Similarly, the input reference points are only used for comparison purposes.
KBPMs have also proven their usefulness for the multispecies management objectives. More precisely, KBPM approach can be applied to analyze the dynamic of the total aggregated biomass and catch of all targeted fish species in a community defining in this way a simple data-limited ecosystem model to assess the ecosystem status (see example).
}
\examples{

library(knobi)

# First step, getting the data from the ICES package
# install.packages("icesSAG")
library(icesSAG)

summary_data <- getSAG(stock = "Hake", year = 2021)
Database <- subset(summary_data, summary_data[,17] == "hke.27.3a46-8abd")
Database <- Database[-nrow(Database),]


# Then, create the data object.

data<-list()
data$Spawning_Biomass<-Database$SSB # We take the SSB in our Database.
data$Catch<-Database$catches # We take the catch in our Database.
data$F_input<-Database$F # We take the F in our Database.
# Reference points estimates from ICES stock assessment model:
# ICES. 2021. Working Group for the Bay of Biscay and the Iberian Waters Ecoregion
# (WGBIE). ICES Scientific Reports. 3:48.1101 pp.
data$RP<-list(F_MSY=0.259, B_MSY=207398, MSY=75052, K=NA)
# In this case, B_MSY is equal to SSB_MSY, since control$method<-"SSB"
# (see control list below).
data$classF_input<-"average" # Character indicating the type of F.
data$years<-Database$Year    # Years corresponding to the catch values
# (can be different than the years corresponding to SSB or biomass).


# Now we define the control.

control<-list()
control$pella<-"TRUE" # Logical. TRUE means that Pella-Tomlinson model is used.
                         # FALSE means that Schaefer model is employed.
control$method<-"SSB" # Information for the fit: "SSB" or "Biomass".


# Finally, we can fit the model
knobi_results<-knobi_fit(data,control,plot_out=TRUE,plot_filename="results")
knobi_results$fit


# Fitting multispecific KBPM

# Below, a multistock approximation aggregating the
# northern and southern stocks of sardine is performed.

# Firstly, read southern stock data
sardine1 <- getSAG(stock = "pil.27.8c9a", year = 2021)
sardine1 <- sardine1[1:44,]
sardine1 <- sardine1[,c(1,6,8,12)]

# Secondly, read northern stock data
sardine2 <- getSAG(stock = "pil.27.8abd", year = 2021)
sardine2 <- sardine2[-nrow(sardine2),]
sardine2 <- sardine2[,c(1,6,8,12)]

# Extract common years of data in both stocks
index <- which(sardine1$Year \%in\% sardine2$Year)
sardine1 <- sardine1[index,]

# Create a data.frame where the SSB and the catch are
# the sum of such data in the two stocks
years<-sardine1$Year
sardine <- data.frame(years=years,SSB=sardine1$SSB+sardine2$SSB,
                      catch=sardine1$catches+sardine2$catches)

# Once the total SSB and catch are available
# we follow previous KBPM illustration
data<-list()
data$Spawning_Biomass<-sardine$SSB
data$Catch<-sardine$catch
data$years<-sardine$years

control<-list(pella="TRUE")

knobi_results2<-knobi_fit(data,control)
knobi_results2$fit



}
\references{
MacCall, A. (2002). Use of Known-Biomass Production Models to Determine Productivity of West Coast Groundfish Stocks. North American Journal of Fisheries Management, 22, 272-279.
}
\author{
\itemize{
\item{Anxo Paz}
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{Maria Grazia Pennino}
}
}
