% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knobi_proj.R
\name{knobi_proj}
\alias{knobi_proj}
\title{KBPM projections}
\arguments{
\item{knobi_results}{The output object of \code{\link{knobi_fit}} function (main package function).}

\item{env_results}{Optional. The output object of \code{\link{knobi_env}} function. The environmental KBPM fit is required to forecast considering the environmental variable(s) values in the future.}

\item{Ct}{Optional. Vector, data frame or matrix that establishes the catch values for each one of the projected years. Different catch scenarios are allowed and can be defined in each of the data frame or matrix columns. Therefore, the length vector (in the case of one scenario) or the number of rows of the data frame or matrix (in the case of multiple scenarios) must be equal to the number of projected years. Projections can be carried out based on a set of future catch values or fishing mortality values, so only one of two arguments, 'Ct' or 'f', is required.}

\item{f}{Optional. Vector, data frame or matrix that establishes the fishing mortality values for each one of the projected years. Different fishing mortality scenarios are allowed and can be defined in each of the data.frame or matrix columns. Therefore, the length vector (in the case of one scenario) or the number of rows of the data frame or matrix (in the case of multiple scenarios) must be equal to the number of projected years. Projections can be carried out based on a set of future catch values or fishing mortality values, so only one of two arguments, 'Ct' or 'f', is required.}

\item{env}{Optional. Environmental variable(s) projections required if the environmental fit is considered to forecast the population and fishery dynamics. This fit considers the variable(s) selected in the \code{\link{knobi_env}} function. If the 'multicovar' argument of \code{\link{knobi_env}} is FALSE, tihs argument is a vector, data frame or matrix containing the values of the environmental covariates (unstandardized) for the projection years (rows) and the different catch or fishing mortality settings (columns).  On the other hand, if the 'multicovar' argument of \code{\link{knobi_env}} is TRUE, the current argument must be a list, and each entry must be a data frame or matrix corresponding to each catch or fishing mortality setting containing the values of the environmental covariates for that scenario.}

\item{plot_out}{Logical. TRUE means that a file with the plot of the forecasts is created. The default value is the input in the \code{\link{knobi_fit}} function.}

\item{plot_dir}{Optional. Directory to create the folder and save the plots. Required when 'plot_out=TRUE'. The default value is the input in the \code{\link{knobi_fit}} function.}

\item{plot_filename}{Optional. Name of the folder that will contain the plots. Required when 'plot_out=TRUE'. The default value is the input in the \code{\link{knobi_fit}} function.}
}
\value{
A list containing the projection results. \itemize{
\item base: three-dimensional matrix containing the projected time series derived from the KBPM base model (see details in \code{link{knobi_fit}}) for each of the catch or fishing mortality scenarios. The first dimension (rows) corresponds to the years, the second dimension (columns) represent the stock quantities (biomass or SSB, surplus production, F and catches) and the third dimension refers to the projection scenarios.
\item add: four-dimensional matrix containing the projected time series derived from the KBPM additive model (see \code{\link{knobi_env}} details) for each of the catch (or fishing mortality) scenarios combined with each of the different environment settings. The first dimension (rows) corresponds to the years, the second dimension (columns) concern the stock quantities (biomass or SSB, surplus production, F and catches), the third dimension refers to the environmental projection scenarios and the fourth dimension corresponds to the catch (or fishing mortality) scenarios. It is only returned if the arguments 'env_results' and 'env' are provided.
\item mult: four-dimensional matrix containing the projected time series derived from the KBPM multiplicative model (see \code{\link{knobi_env}} details) for each of the catch (or fishing mortality) scenarios combined with each of the different environment settings. The first dimension (rows) corresponds to the years, the second dimension (columns) concern the stock quantities (biomass or SSB, surplus production, F and catches), the third dimension refers to the environmental projection scenarios and the fourth dimension corresponds to the catch (or fishing mortality) scenarios. It is only returned if the arguments 'env_results' and 'env' are provided.
\item df: data frame containing historical and projected biomass (or SSB), catch, SP and fishing mortality values for the different scenarios.
\item plots: data frame containing the plots with the projections, for each scenario or combination of scenarios and for each model.}
The resulting plots are displayed in the plot window and are also saved (if plot_out="TRUE") in the  provided directory or in the same directory as \code{link{knobi_fit}}.
If only the base KBPM is considered in the projections, a plot is presented with a panel reporting the catch, fishing mortality, SP and biomass (or SSB) for each catch or fishing mortality scenario.
Furthermore, if the environmental fit is also considered, the plots are displayed with panels for each catch or fishing mortality scenario and for each environmental scenario.
}
\description{
Projection of future population and fishery dynamics is carried out for a given set of management targets. More precisely, the function projects the time series of biomass (or spawning biomass) and then the surplus production for a set of future catch or fishing mortality values.
}
\examples{

\dontrun{

### Projecting through catch with no environmental information

# Then, create the data frame containing the selected catch for the projected
# years. In this illustration, within each scenario, the catch values are
# constant through the projected years. Three scenarios are considered:
# (i) catch value equal to the last historical catch multiplied by 1,
# (ii) last historical catch multiplied by 1.2 and
# (iii) last historical catch multiplied by 0.8.

catch<-rep(knobi_results$df$C[length(knobi_results$df$C)],5)

Ct<-data.frame(catch=catch,
              catch08=0.8*catch,
              catch12=1.2*catch)

# Then, knobi_proj function can be applied

knobi_proj(knobi_results, Ct=Ct)


### With environmental information

# In this case, in addition to the previous example, the 'knobi_env' example
# has to be run at first

# We include the future values of the environmental variable(s) in a data
# frame containing the environmental covariable values for the projected
# years. Three scenarios are considered:
# (i) Constant maximum year temperature equal to 19,
# (ii) Temperature equal to 19 and then constant raise of 0.5 degrees
# (iii) Temperature equal to 19 and then constant raise of 1 degrees

env<-data.frame(Tmax_cte=c(19,19,19,19,19),
               Tmax_05=c(19,19.5,20,20.5,21),
               Tmax_1=c(19,20,21,22,23))

# Based on the previous objects we can apply the projection function.

knobi_proj(knobi_results, knobi_environmental, Ct=Ct, env=env)


### Through fishing mortality without environmental information

# Alternatively, projections can be based on fishing mortality.
# The scenarios presented below have been created from the estimated F_msy of
# knobi_fit analysis.

fmsy<-knobi_results$BRPs['F_MSY']
ff<-rep(fmsy,8)
f<-data.frame(f=ff,f12=ff*1.2,f08=ff*0.8)

knobi_proj(knobi_results, f=f)


### Through fishing mortality with environmental information

knobi_proj(knobi_results, f=f[1:5,], env_results=env_results, env=env)


# In case of multicovar<-TRUE in knobi_env, a list is required in which
# each item is a data frame for each environmental scenario

env<-list(climate_1=data.frame(AMO=c(0.2,0.2,0.3,0.3,0.4),
                              Tmax=c(19,19,20,20,21)),
          climate_2=data.frame(AMO=c(0.2,0.3,0.4,0.5,0.6),
                              Tmax=c(19,20,21,22,23)))

knobi_proj(knobi_results, knobi_environmental2, Ct=Ct, env=env)
}

}
\author{
\itemize{
\item{Anxo Paz}
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{M. Grazia Pennino}
}
}
