% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knobi_retro.R
\name{knobi_retro}
\alias{knobi_retro}
\title{KBPM retrospective analysis}
\usage{
knobi_retro(
  knobi_results,
  yR = NULL,
  yR0 = NULL,
  nR = NULL,
  plot_out = F,
  plot_filename = NULL,
  plot_dir = NULL
)
}
\arguments{
\item{knobi_results}{A list containing the results of the KBPM fit. Object provided by \code{\link{knobi_fit}} function (main package function).}

\item{yR}{Vector of catch time series final years in each one of the retrospective models.}

\item{yR0}{Optional. Vector of catch time series starting years in each one of the retrospective models. The same length of 'yR' vector is required. By default, the catch time series is assumed to start in the same year as in the original fit.}

\item{nR}{Number of retrospective patterns. Only required when 'yR' is not provided (if both arguments are included and error message is reported). See details.}

\item{plot_out}{Logical. TRUE means that a file with the plot of the retrospective fits is created. The default value is the input in the \code{\link{knobi_fit}} function.}

\item{plot_filename}{Optional. Name of the folder that will contain the plots. Required when 'plot_out=TRUE'. The default value is the input in the \code{\link{knobi_fit}} function.}

\item{plot_dir}{Optional. Directory to create the folder and save the plots. Required when 'plot_out=TRUE'. The default value is the input in the \code{\link{knobi_fit}} function.}
}
\value{
A list containing the retrospective analysis: parameter estimates and reference points for each one of the models.
The estimated surplus production curves from the retrospective analysis are plotted. The plot is displayed in the plot window and saved (if plot_out=TRUE) in the provided directory or in the current directory.
}
\description{
This function performs a retrospective analysis that evaluates the robustness of the KBPM fit to the systematically deletion of recent data.
}
\details{
There are different options for defining retrospective fits:
(1) Usage of 'nR' argument. This argument specifies the number of retrospective patterns. Furthermore, it is implicit in the use of this argument that the retrospective patterns will consist of the systematic deletion of the last year of data up to the number of years determined by 'nR'.
(2) Usage of 'yR' argument. This argument specifies the catch time series final years in each one of the retrospective models, which allows greater flexibility by being able to choose the year from which we delete information. Then, the number of retrospective patterns will be the length of the vector 'yR'. In this option it is possible to set different starting years for the respective time series final years through the argument 'yR0'.
}
\examples{

\dontrun{

library(knobi)

# See knobi_fit example to obtain the knobi_results object
knobi_retrospectives<-knobi_retro(knobi_results,nR=5,plot_out=T)
knobi_retrospectives

knobi_retro(knobi_results,yR=c(2010,2015),yR0=c(1995,2000))
}

}
\author{
\itemize{
\item{Anxo Paz}
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{M. Grazia Pennino}
}
}
