% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.knobi.R
\name{print.knobi}
\alias{print.knobi}
\title{Print a knobi object}
\usage{
\method{print}{knobi}(x, ...)
}
\arguments{
\item{x, ...}{Fitted model objects of class \code{knobi} produced by \code{knobi_fit()}, \code{knobi_env()} or \code{knobi_proj()}.}
}
\description{
The default print method for a \code{\link{knobi_fit}}, \code{\link{knobi_env}} or a \code{\link{knobi_proj}} object
}
\details{
Prints out the formula and the parameters estimates for the base KBPM fit or the environmental KBPM fit or the biomass and the surplus production estimated projections.
}
\seealso{
\code{\link{knobi_fit}}, \code{\link{knobi_env}}, \code{\link{knobi_proj}}
}
\author{
\itemize{
\item{Anxo Paz}
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{M. Grazia Pennino}
}
}
